+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Kernelised Stein Discrepancy divergence in `coreax.metrics.KSD`.
 - Added Conditional Maximum Mean Discrepancy (CMMD) metric in `coreax.metrics.CMMD`
 - Added capability to take tensor products of kernels in `coreax.kernel.TensorProductKernel`
-
+- Added Joint Maximum Mean Discrepancy (JMMD) metric in `coreax.metrics.JMMD`
 
 ### Fixed